requirements.txt
numpy>=1.10.4
wikipedia>=1.4.0
requests>=2.8.1
scipy>=0.16.0
nose>=1.3.7
theano>=0.8.2
nltk>=3.0.0
scriptine>=0.2.1
mxnet-cu90>=1.2.1
seaborn>=0.9.0
subprocess32>=3.2.7
matplotlib>=2.1.2
